refactor: modularize rateProvider to make it easier to test
